@@ -1,3 +1,3 @@
 import { ModuleWithProviders, Type } from '@angular/core';
 
-export type AngularModule = Type<any> | ModuleWithProviders;
+export type AngularModule = Type<any> | ModuleWithProviders<any>;
