@@ -1,17 +1,16 @@
 import { Component, Input, NgModule, Output, EventEmitter } from '@angular/core';
 import { Shallow } from '../shallow';
-import { ComponentFixtureAutoDetect } from '@angular/core/testing';
 
 ////// Module Setup //////
 @Component({
   selector: 'list-container',
-  template: '<ul><ng-content></ng-content></ul>'
+  template: '<ul><ng-content></ng-content></ul>',
 })
 class ListContainerComponent {}
 
 @Component({
   selector: 'list-item',
-  template: '<li [class.bold]="bold" (click)="select.emit(true)"><ng-content></ng-content></li>'
+  template: '<li [class.bold]="bold" (click)="select.emit(true)"><ng-content></ng-content></li>',
 })
 class ListItemComponent {
   @Input() bold = false;
@@ -27,7 +26,7 @@ class ListItemComponent {
       <list-item id="chuck" [bold]="true" (select)="chuckSelected($event)">Chuck Norris</list-item>
       <list-item>Tom Hanks</list-item>
     </list-container>
-  `
+  `,
 })
 class AwesomeListComponent {
   @Input() topItem!: string;
@@ -39,7 +38,7 @@ class AwesomeListComponent {
 }
 
 @NgModule({
-  declarations: [ListContainerComponent, ListItemComponent, AwesomeListComponent]
+  declarations: [ListContainerComponent, ListItemComponent, AwesomeListComponent],
 })
 class ListModule {}
 //////////////////////////
@@ -55,7 +54,10 @@ describe('multiple components', () => {
     const { find } = await shallow.render('<awesome-list></awesome-list>');
 
     // Note we query by the component here
-    expect(find(ListItemComponent).map(li => li.nativeElement.innerText.trim())).toEqual(['Chuck Norris', 'Tom Hanks']);
+    expect(find(ListItemComponent).map((li) => li.nativeElement.innerText.trim())).toEqual([
+      'Chuck Norris',
+      'Tom Hanks',
+    ]);
   });
 
   it('reports when Chuck is pressed', async () => {
